Merge nginx.conf file for different distros for 1.12 and 1.14
Users (or a test script) need to schuffle with config files to make
the example app working, which is not acceptable user experience.

This change results in just one copy of the nginx.conf config file in
the test-app, while using the non-SCL paths. For making it work, we
need to add compatibility symlinks for directories used in this config
file.

Particularly, /usr/share/nginx and /etc/nginx will point to respective
SCL directories, and /var/run needs to be writable the same as in RHEL-8
case. Those changes should be back-ward compatible, because if anybody
used SCL paths before, they are still there, just symlinks point to
those as well.